32 persistencia de la entidad repartidor

+++ src/main/java/com/progra3/proyecto/entity/Vehiculo.java
@@ -10,17 +10,19 @@ public class Vehiculo {
 	 
 		@Id
 		@GeneratedValue(strategy = GenerationType.IDENTITY)	
-		private Long id;
+		private int id;
 		private String tipo;		
 		private String nombre;		
 		private String color;
 
-		public Long getId() {
+		public int getId() {
 			return id;
 		}
-		public void setId(Long id) {
+
+		public void setId(int id) {
 			this.id = id;
 		}
+
 		public String getTipo() {
 			return tipo;
 		}
